requirements.txt
django-netjsonconfig>=0.7.1,<0.8.0
openwisp-utils[users]<0.2
# geo
#TODO: pin versions
django-leaflet
djangorestframework-gis
channels
